test: fix more of NonlinearSolve tests
