+++ .github/workflows/CI_BracketingNonlinearSolve.yml
@@ -26,7 +26,7 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - "lts"
+          - "1.10"
           - "1"
         os:
           - ubuntu-latest
