+++ .github/workflows/CI_NonlinearSolveBase.yml
@@ -25,7 +25,7 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - "lts"
+          - "1.10"
           - "1"
         os:
           - ubuntu-latest
