+++ .github/workflows/CI_SciMLJacobianOperators.yml
@@ -24,7 +24,7 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - "lts"
+          - "1.10"
           - "1"
         os:
           - ubuntu-latest
