+++ Project.toml
@@ -78,7 +78,7 @@ InteractiveUtils = "<0.0.1, 1"
 LeastSquaresOptim = "0.8.5"
 LineSearch = "0.1.4"
 LineSearches = "7.3"
-LinearAlgebra = "1.11.0"
+LinearAlgebra = "1.10"
 LinearSolve = "2.36.1"
 MINPACK = "1.2"
 MPI = "0.20.22"
